The aws-ecr-push role fails with multiple images
This should fix it so that it won't fail on the second image.

It should also fix the case where new registries cause a failure.

This does change the expected type for aws_ecr_push_image to a list,
there isn't a simple way to do that in Jinja and we don't want to fork
the ecs_ecr module any more than we already have.