Fix ecr_retag from aws-ecr-promote
This module had a few bugs that weren't apparent at first.

First, the
image refs in the 'repos' item are expected to be image refs for
Kubernetes to pull, so they can't start with https://.

Second, the ref was being pulled after it had the image name stripped
off, resulting in double-plus-ungood broken image refs.